pyproject.toml
[tool.poetry]
name = "e2e-tests"
version = "0.1.0"
description = ""
authors = ["Dani Santamaría <daniel.santamaria@innovamat.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
testcontainers = "^3.7.1"
flask-sqlalchemy = "^3.1.1"
pymysql = "^1.1.0"
mysql-connector-python = "^8.2.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
